+++ src/CharacterBase/CharacterBase.cpp
