+++ src/Easing/easing.cpp
