+++ src/Player/Player.h
@@ -16,7 +16,6 @@ public:
 	Vec2f getSize();
 
 private:
-
 	void move();
 
 };
